refactor work experience to support multiple positions

+++ src/modules/work/handlers/delete-experience-by-id.handler.ts
@@ -11,28 +11,46 @@ export const deleteExperienceById = factory.createHandlers(
     "param",
     z.object({
       id: z.string(),
+      positionId: z.string(),
     }),
   ),
+
   async (c) => {
     try {
-      const { id } = c.req.valid("param");
-      const experience = await WorkExperienceModel.findByIdAndDelete(id);
+      const { id, positionId } = c.req.valid("param");
 
-      // if experience does not exists
+      const experience = await WorkExperienceModel.findById(id);
       if (!experience) {
-        throw new HTTPException(404, {
+        throw new HTTPException(StatusCodes.HTTP_404_NOT_FOUND, {
           message: "Experience not found",
         });
       }
 
-      return c.json({
-        message: "experience deleted",
-        experience,
-      });
-    } catch (err) {
-      if (err instanceof HTTPException) {
-        throw err;
+      const position = experience.positions.id(positionId);
+      if (!position) {
+        throw new HTTPException(StatusCodes.HTTP_400_BAD_REQUEST, {
+          message: `Position with id ${positionId} does not exist`,
+        });
       }
+
+      position.deleteOne();
+
+      let result;
+      if (experience.positions.length === 0) {
+        result = await WorkExperienceModel.findByIdAndDelete(id);
+      } else {
+        result = await experience.save();
+      }
+
+      return c.json(
+        {
+          message: "Position deleted successfully",
+          data: result,
+        },
+        StatusCodes.HTTP_200_OK,
+      );
+    } catch (err) {
+      if (err instanceof HTTPException) throw err;
 
       logger.error("Error while deleting work experience", {
         module: "work",
